refactor: use memmap in hello world (#2584)
* refactor: use memmap in hello world

* refactor: use memmap in hello world

* refactor: use memmap in hello world

* test: fix broken test
