+++ tests/system/fashion/test_fashion.py
@@ -4,7 +4,7 @@ import pytest
 
 import jina
 from jina.helloworld.fashion.app import hello_world
-from jina.helloworld.fashion.executors import *
+from jina.helloworld.fashion.my_executors import *
 from jina.parsers.helloworld import set_hw_parser
